Lesson_3_4 homework in process

+++ main.py
@@ -1,5 +1,5 @@
 #Created Homework branch for the main file
-num1 = int(input("Enter first number: "))
-num2 = int(input("Enter second number: "))
-num3 = int(input("Enter third number: "))
+num1 = float(input("Enter first number: "))
+num2 = float(input("Enter second number: "))
+num3 = float(input("Enter third number: "))
 print(f"Result: {(num1 > 0)+(num2 > 0)+(num3 > 0)} number(s) are positive")
